# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "My Cat and dog server api swagger",
      "description": "Yet Another Node.js Cat and dog server api swagger",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths":{
...
    "/person":{
        "get" :{
            "tags":["Persons"],
            "summary": "Get all Persons",
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Persons"
                    }
                }
            }
        },
        "post":{
            "tags":["Persons"],
            "summary": "Create a new  person",
            "parameters":[
                {
                    "in":"body",
                    "name":"body",
                    "description":"The person we want to add ",
                    "schema":{
                        "$ref": "#definitions/Person"
                    }
                }
            ],
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Persons"
                    }
                }
            }
        }
    },
    "/person/{id}":{
        "parameters":[
            {
                "in":"path",
                "name":"id",
                "required":true,
                "description":"ID of the person that we want to match ",
                "type":"string"
            }
        ],
        "get": {
            "tags":["Persons"],
            "summery": "get person with given id",
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Person with id",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ], 
            
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Person"
                    }
                }
            }
        },
        "put": {
            "summary": "Update person with given ID",
            "tags": [
                "Persons"
            ],
            "requestBody": {
                "description": "Person Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                }
            },
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Person with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                },
                {
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "Person with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/Person"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/Person"
                    }
                },
                "400": {
                    "description": "Failed. Bad post data."
                },
                "404": {
                    "description": "Failed. Cat not found."
                }
            }
        },
        "delete": {
            "summary": "Delete Person with given ID",
            "tags": [
                "Persons"
            ],
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Delete Person with id",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                },
                "404": {
                    "description": "Failed. Person not found."
                }
            }
        }
    },
    "/persons":{
        
        "get": {
            "tags":["Persons"],
            "summary": "get persons with given search",
            "parameters":[
                {
                    "in":"body",
                    "name":"body",
                    "description":"The persons we want to find ",
                    "schema":{
                        "$ref": "#definitions/Person"
                    }
                }
            ], 
            
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Persons"
                    }
                }
            }
        }
    },
    "/owner":{
        "get" :{
            "tags":["Owners"],
            "summary": "Get all owners",
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Owners"
                    }
                }
            }
        },
        "post":{
            "tags":["Owners"],
            "summary": "Create a new  owner",
            "parameters":[
                {
                    "in":"body",
                    "name":"body",
                    "description":"The owner we want to add ",
                    "schema":{
                        "$ref": "#definitions/Owner"
                    }
                }
            ],
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Owners"
                    }
                }
            }
        }
    },
    "/owner/{id}":{
        "parameters":[
            {
                "in":"path",
                "name":"id",
                "required":true,
                "description":"ID of the owner that we want to match ",
                "type":"string"
            }
        ],
        "get": {
            "tags":["Owners"],
            "summary": "get owner with given id",
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "owner with id",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ], 
            
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#/definitions/Owner"
                    }
                }
            }
        },
        "put": {
            "summary": "Update owner with given ID",
            "tags": ["Owners"],
            "requestBody": {
                "description": "Person Object",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                }
            },
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Owner  with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                },
                
                {
                    "in": "body",
                    "name": "body",
                    "required": true,
                    "description": "Owner with new values of properties",
                    "schema": {
                        "$ref": "#/definitions/Owner"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/owner"
                    }
                },
                "400": {
                    "description": "Failed. Bad post data."
                },
                "404": {
                    "description": "Failed. Cat not found."
                }
            }
        },
        "delete": {
            "summary": "Delete owner with given ID",
            "tags": [
                "Owners"
            ],
            "parameters": [
                {
                    "in": "path",
                    "name": "id",
                    "required": true,
                    "description": "Delete owner with id",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "$ref": "#/definitions/id"
                    }
                },
                "404": {
                    "description": "Failed. owner not found."
                }
            }
        }
    },
    "/owners":{
        
        "get": {
            "tags":["Owners"],
            "summary": "get owners with given search",
            "parameters":[
                {
                    "in":"body",
                    "name":"body",
                    "description":"The owners we want to find ",
                    "schema":{
                        "$ref": "#definitions/Owner"
                    }
                }
            ], 
            
            "products":["application/json"],
            "responses":{
                "200":{
                    "description":"ok",
                    "schima":{
                        "$ref": "#definitions/Owners"
                    }
                }
            }
        }
    }
...
    },
    "definitions":{
        "id":{
            "properties":{
                "uuid":{
                    "type": "string"
                }
            }
        },
...
        "Person":{
            "type" : "object",
            "properties":{
                "firstName": { "type": "string"},
                "lastName": { "type": "string"},
                "phoneNumber": {"type" : "integer"},
                "address": { "type": "string"}              
            }
        },
...
        "Owner":{
            "type" : "object",
            "properties":{
                "personId": { "type": "string"},
                "catId": { "type": "string"},
                "dogId": { "type": "string"} 
                
            }
        },
...
    }
}